Rename sortitionpools.ts to operators.ts
The file no longer works with sortitionpool only. It uses random beacon
to register operator.
