+++ solidity/random-beacon/test/utils/groups.ts
@@ -2,7 +2,7 @@ import { helpers } from "hardhat"
 import { signAndSubmitDkgResult } from "./dkg"
 import { constants, params } from "../fixtures"
 import blsData from "../data/bls"
-import { Operator } from "./sortitionpool"
+import { Operator } from "./operators"
 import type { RandomBeacon } from "../../typechain"
 
 const { mineBlocks } = helpers.time
